synchronization merge: use new merge if available
Git patch "merge-recursive: option to disable renames" added support for
disabling rename detection with the recursive strategy. If not
available, we fall back to the resolve strategy, which is safe but does
not support custom merge drivers.

This conditional dispatch logic is handled by the new script
utils/git_merge_no_renames