Refactor mirrored repository layout
As part of the work to allow a mirrored repository to be used as a
working one, it is convenient to redesign its layout. We make the data
repository the top folder and move everything else into an untracked
folder .dow

A script dow-convert is included for porting old repositories to the new
layout.

Previous layout:
  top.dow/data
  top.dow/versions
  top.dow/map
  top.dow/[etc.]

New layout:
  top/
  top/.dow/versions
  top/.dow/map
  top/.dow/[etc.]