dow-put: support unsynchronized branches
Branches in the data repository do not need to be synchronized anymore.
Instead synchronization is done by calling dow-put.

Branches prefixed with "sync/" in the data repository are introduced to
track the corresponding branches in the versions repository. So, for
instance, "sync/master" represents the latest synchronized commit from
master, which is mapped to "master" in the versions repository.

Calling dow-put on an old mirrored repository automatically creates the
"sync/" corresponding to the current branch, so that no further setup is
required.

The functionality of utils/git_pull is now equivalent to a git-fetch
followed by a dow-put, so it was discontinued.